pyproject.toml
[project]
name = "aixbio"
version = "0.1.0"
description = "Digital-to-biological pipeline for biocompound replication"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.87",
    "langgraph>=0.2",
    "langchain-openai>=0.3",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "viennarna>=2.7.2",
]

[project.scripts]
aixbio = "aixbio.__main__:main"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
